requirements.txt
requests
Pillow
editdistance
